+++ Assets/OxGKit/Utilities/Scripts/Runtime/Timer/DTUpdater.cs
@@ -96,6 +96,11 @@ namespace OxGKit.Utilities.Timer
             this._cts = null;
         }
 
+        public bool IsRunning()
+        {
+            return this._isRuning;
+        }
+
         private async UniTaskVoid _SetInterval(CancellationTokenSource cts)
         {
             try