+++ Assets/OxGKit/Utilities/Scripts/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [1.3.0] - 2025-03-10
 - Added onLateUpdate callback in DTUpdater and RTUpdater.
+- Added IsRunning() method in DTUpdater, RTUpdater, IntervalTimer.
+- Added CheckIsRunning() method in IntervalSetter.
 
 ## [1.2.2] - 2025-02-11
 - Added delayFrameAfterSpawnCount for NodePool (Allows setting a delay after producing N objects, where a delay of N frames occurs after every N objects are produced).